dynamic-slides/requirements.txt
# fal.ai API client
fal-client>=0.5.0

# Environment variables
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0

# HTTP requests
httpx>=0.25.0

# Progress bar
tqdm>=4.66.0
